requirements.txt
# Vero Capital Holdings — Real Estate Investment Pipeline
# Phase 1: Data Ingestion Dependencies

# Core data stack
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0

# HTTP & API
requests>=2.31.0
requests-cache>=1.1.0
ratelimiter>=1.2.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
responses>=0.24.0

# Data validation
pandera>=0.17.0
